+++ AltairCA.EntityFrameworkCore.PostgreSQL.ColumnEncryption/Attribute/AttributeEncryptExtension.cs
@@ -6,13 +6,13 @@ namespace AltairCA.EntityFrameworkCore.PostgreSQL.ColumnEncryption.Attribute
 {
     internal static class AttributeEncryptExtension
     {
-        public static string NpgsqlEncrypt(this string value, string password,string iv,int keyLength)
+        internal static string NpgsqlEncrypt(this string value, string password,string iv,int keyLength)
         {
             if (string.IsNullOrEmpty(value))
                 return null;
             return AesUtil.AES_encrypt(value, password,iv,keyLength);
         }
-        public static string NpgsqlDecrypt(this string value, string password,string iv,int keyLength)
+        internal static string NpgsqlDecrypt(this string value, string password,string iv,int keyLength)
         {
             if (string.IsNullOrEmpty(value))
                 return null;