+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Unreleased
+
+- Added `UseEncryption` extension to use with property configuration.
+
 ### 6.0.2.2
 
 - Fixed bug with working with multi db contexts with different passwords ( https://github.com/XHunter74/EntityFrameworkCore.PostgreSQL.ColumnEncryption/tree/v6.0.2 ). Special thanks to Serhiy Krasovskyy (https://github.com/XHunter74)